requirements.txt
flask==3.0.3
gspread==6.1.2
google-auth==2.29.0
python-dotenv==1.0.1
google-cloud-bigquery==3.25.0
gunicorn==22.0.0
openpyxl==3.1.5
psycopg2-binary==2.9.9
